+++ app/TestScreen.tsx
@@ -1,0 +1,15 @@
+import React from 'react'
+import { View } from 'react-native';
+import { NavigationComponent } from 'react-navigation';
+
+interface IProps {
+    navigation: NavigationComponent
+}
+
+export default class TestScreen extends React.Component<IProps> {
+    render() {
+        return (
+            <View />
+        )
+    }
+}