+++ app/TodoInput.tsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import { View, AppRegistry, TextInput, Text, Button } from 'react-native';
+import { View, AppRegistry, TextInput, Text, Button, StyleSheet } from 'react-native';
 import { observable, computed, action } from 'mobx';
 import { observer } from 'mobx-react';
 import { todoStore } from './todoStore'
@@ -13,23 +13,23 @@ export class TodoInput extends Component {
         this.onPress = this.onPress.bind(this)
     }
 
-    @observable titleInput: string
+    @observable titleInput = ""
 
     @action onChange(event) {
         this.titleInput = event.nativeEvent.text
     }
 
     @action onPress() {
-        todoStore.addTodo(this.titleInput, false)
-        this.titleInput = ""
+        if(this.titleInput != "") {
+            todoStore.addTodo(this.titleInput, false)
+            this.titleInput = ""
+        }
     }
 
     render() {
         return(
-            <View style={{width: "50%"}}>
-                <View style={{
-                    backgroundColor: "grey"
-                }}>
+            <View style={styles.root}>
+                <View style={styles.inputView}>
                     <TextInput
                         placeholder="Todo title"
                         value={this.titleInput}
@@ -37,7 +37,7 @@ export class TodoInput extends Component {
                     />
                 </View>
 
-                <View style={{paddingVertical: 5}} />
+                <View style={styles.divider} />
 
                 <View>
                     <Button
@@ -50,5 +50,17 @@ export class TodoInput extends Component {
         )
     }
 }
+
+const styles = StyleSheet.create({
+    root: {
+        width: "50%"
+    },
+    inputView: {
+        backgroundColor: "grey"
+    },
+    divider: {
+        paddingVertical: 5
+    }
+})
 
 AppRegistry.registerComponent("ReactNativeTest", () => TodoInput)